Write tests for installation webhook and one for comment
